+++ batch-example/src/main/java/demo/model/SourceData.java
@@ -2,17 +2,17 @@ package demo.model;
 
 public class SourceData {
 
-	private Integer value;
+	private Long value;
 
-	public SourceData(Integer value){
+	public SourceData(Long value){
 		this.value = value;
 	}
 
-	public Integer getValue() {
+	public Long getValue() {
 		return value;
 	}
 
-	public void setValue(Integer value) {
+	public void setValue(Long value) {
 		this.value = value;
 	}
